Move Tyvar & Tycon up a level in folder structure
