+++ lib/Besra/Parser/Tyvar.hs
@@ -2,7 +2,7 @@ module Besra.Parser.Tyvar ( parser ) where
 
 import Protolude
 import Besra.Parser.Helpers
-import Besra.Types.IR1.Tyvar
+import Besra.Types.Tyvar
 import Besra.Types.Id
 import Besra.Types.Ann
